feat(serve): add check-in management functionality
- Create CheckIn domain entity with properties for resident information and check-in details
- Implement CheckInConfig domain for managing check-in configurations and fees
- Add CheckInApplyDto for handling check-in application requests
- Create CheckInContractDto for contract information during check-in process
- Implement CheckInElderDto and CheckInElderVo for elder resident data transfer
- Add CheckInConfigDto and CheckInConfigVo for configuration data handling
- Create CheckInDetailVo for comprehensive check-in detail responses
- Implement CheckInController with full CRUD operations for check-in management
- Add CheckInConfigController with complete REST API endpoints
- Create CheckInMapper and CheckInConfigMapper with MyBatis operations
- Implement CheckInService and CheckInConfigService business logic
- Add MyBatis XML mappers for database operations
- Integrate Swagger API documentation annotations
- Add Excel export functionality for check-in data

+++ YiSmartCloud-background/YiSmartCloud-nursing-platform/src/main/java/org/FlyingSparrow/YiSmartCloud/serve/domain/Room.java
@@ -36,7 +36,7 @@ public class Room extends BaseEntity {
 
     /** 房间类型: 1-单人间, 2-双人间, 3-多人间 */
     @Excel(name = "房间类型: 1-单人间, 2-双人间, 3-多人间")
-    private Long roomType;
+    private Integer roomType;
 
     /** 房间介绍 */
     @Excel(name = "房间介绍")
